modif

+++ VM/champions_maker.c
@@ -122,7 +122,8 @@ int		champion_validity_checker(t_var *var, t_champ *champ)
 		return(0);
 	}
 	champ->name = ft_hextoa(&stock[EXEC_MAGIC_LENGTH / 2], PROG_NAME_LENGTH);
-	if ((champ->size = stock[PROG_NAME_LENGTH / 2 + EXEC_MAGIC_LENGTH]) != (file_size - COR_MIN_VALUE) / 2)
+	if ((champ->size = stock[PROG_NAME_LENGTH / 2 + EXEC_MAGIC_LENGTH]) !=
+		(unsigned int)((file_size - COR_MIN_VALUE) / 2))
 	{
 		error_manager(-5, var);
 		return (0);
@@ -160,6 +161,6 @@ int		champions_maker(t_dbllist *champ_list, t_options *options, t_var *var)
 	if (set_champion(var, options, &champ) == 0)
 		return (0);
 	ft_lstdbladd_tail(champ_list, &champ, sizeof(t_champ));
-	
+
 	return (1);
 }
